#1033: unified cryptoTool usage by carina
