+++ carina-utils/src/main/java/com/qaprosoft/carina/core/foundation/utils/R.java
@@ -19,14 +19,13 @@ import java.net.URL;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Properties;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import com.qaprosoft.carina.core.foundation.crypto.CryptoTool;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.log4j.Logger;
 
 import com.qaprosoft.carina.core.foundation.commons.SpecialKeywords;
+import com.qaprosoft.carina.core.foundation.crypto.CryptoTool;
 import com.qaprosoft.carina.core.foundation.exception.InvalidConfigurationException;
 
 /**
@@ -161,7 +160,7 @@ public enum R {
     /**
      * Returns value either from systems properties or config properties context.
      * Systems properties have higher priority.
-     * Decryption is performed if required.
+     * Decryption is not performed!
      * 
      * @param key Requested key
      * @return config value
@@ -176,14 +175,22 @@ public enum R {
         value = CONFIG.resourceFile.equals(resourceFile) ? PlaceholderResolver.resolve(propertiesHolder.get(resourceFile), key)
                 : propertiesHolder.get(resourceFile).getProperty(key);
 
+        // [VD] Decryption is prohibited here otherwise we have plain sensitive information in logs! 
 
-        if(isEncrypted(value, CRYPTO_PATTERN)) {
-            value = decrypt(value, CRYPTO_PATTERN);;
-        }
-
-        // TODO: why we return empty instead of null?
         // [VD] as designed empty MUST be returned
         return value != null ? value : StringUtils.EMPTY;
+    }
+    
+    /**
+     * Returns decrypted value either from systems properties or config properties context.
+     * Systems properties have higher priority.
+     * Decryption is performed if required.
+     * 
+     * @param key Requested key
+     * @return config value
+     */
+    public String getDecrypted(String key) {
+        return decrypt(get(key), CRYPTO_PATTERN);
     }
 
 
@@ -248,19 +255,6 @@ public enum R {
         }
         
         return testProperties.get();
-    }
-
-    private boolean isEncrypted(String content, Pattern pattern) {
-        String wildcard = pattern.pattern().substring(pattern.pattern().indexOf("{") + 1,
-                pattern.pattern().indexOf(":"));
-        if (content != null && content.contains(wildcard)) {
-            Matcher matcher = pattern.matcher(content);
-            if (matcher.find()) {
-                LOGGER.debug("'" + content + "' require decryption.");
-                return true;
-            }
-        }
-        return false;
     }
 
     private String decrypt(String content, Pattern pattern) {
