+++ carina-utils/src/test/java/com/qaprosoft/carina/core/utils/RTest.java
@@ -101,18 +101,19 @@ public class RTest {
 
     @Test
     public void testEncryption() {
-        Assert.assertEquals(R.CONFIG.get("password"), "EncryptMe");
+        Assert.assertEquals(R.CONFIG.get("password"), "{crypt:8O9iA4+f3nMzz85szmvKmQ==}");
+        Assert.assertEquals(R.CONFIG.getDecrypted("password"), "EncryptMe");
         //Assert.assertEquals(R.CONFIG.getSecured("password"), "{crypt:8O9iA4+f3nMzz85szmvKmQ==}")
     }
 
     @Test
     public void testPlaceholdersWithEncryption() {
-        Assert.assertEquals(R.CONFIG.get("credentials"), "test@gmail.com/EncryptMe");
+        Assert.assertEquals(R.CONFIG.getDecrypted("credentials"), "test@gmail.com/EncryptMe");
     }
 
     @Test
     public void testPlaceholdersWithEncryptionTestData() {
-        Assert.assertEquals(R.TESTDATA.get("test_credentials"), "test@gmail.com/EncryptMe");
+        Assert.assertEquals(R.TESTDATA.getDecrypted("test_credentials"), "test@gmail.com/EncryptMe");
     }
 
     @Test
